requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.8.0
asyncio
python-dotenv>=1.0.0

# API and web frameworks
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
websocket-client>=1.6.0

# System monitoring
psutil>=5.9.0

# Data compression
zstandard>=0.21.0

# Data and analysis
yfinance>=0.2.28
ta>=0.10.2
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Cloud services
boto3>=1.28.0

# Timezone handling
pytz>=2023.3
python-dateutil>=2.8.0

# Slack integration
slack-sdk>=3.23.0

# Logging and utilities
colorama>=0.4.6
tqdm>=4.65.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0

# JSON handling
orjson>=3.9.0